+++ src/auth/auth-provider.ts
@@ -9,7 +9,7 @@ import fetch from "node-fetch";
 import { v4 as uuid } from "uuid";
 import vscode, { AuthenticationSession, Disposable } from "vscode";
 import { z } from "zod";
-import { AUTHORIZATION_HEADER } from "../colab/headers";
+import { AUTHORIZATION_HEADER } from "../workbench/headers";
 import { Toggleable } from "../common/toggleable";
 import { Credentials } from "./login";
 import { AuthStorage, RefreshableAuthenticationSession } from "./storage";
