+++ src/jupyter/provider.unit.test.ts
@@ -14,7 +14,7 @@ import sinon from "sinon";
 import { SinonStubbedInstance } from "sinon";
 import vscode from "vscode";
 import { GoogleAuthProvider } from "../auth/auth-provider";
-import { WORKBENCH_COMMAND } from "../colab/commands/constants";
+import { WORKBENCH_COMMAND } from "../workbench/constants";
 import { newVsCodeStub, VsCodeStub } from "../test/helpers/vscode";
 import { ProjectsClient } from "../workbench/projects-client";
 import { WorkbenchJupyterServerProvider } from "./provider";
